+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Input consumption now works properly for contexts with different schedules.
 - `Pulse` condition triggered `Fired` only on the first actuation.
 - `HoldAndRelease` condition fired only on exact duration match.